requirements.txt
flake8==7.1.0
huggingface-hub==0.23.4
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.3
langchain-openai==0.1.8
langsmith==0.1.67
llama-index==0.10.46
llama-index-agent-openai==0.2.7
llamaindex-py-client==0.1.19
openai==1.30.5
pydantic==2.6.3
pyflakes==3.2.0
sentence-transformers==3.0.1
weaviate-client==4.6.4
pinecone-client==4.1.2
pinecone-plugin-interface==0.0.7